Fixed a couple of buglets in value range iteration in ArrayIntMap, and added consistency checks there and to IntMapModifier.
Ported same range iteration features to IntSet and implemented for BitSet- and
array-based implementations.

Added Trimmable to allow clients to optimize the footprint of array-based
structures to only the memory required once they reach a stable state.

IntSet javadocs.

Support of finding the end of a contiguous set of numbers in IntSet, to
facilitate its use managing pools of objects stored in an array; added
inverse set and range set support for the same reason.

Tests for all of the above.